requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.4
django-allauth==0.52.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-environ==0.9.0
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
